move \filters into their own file
