+++ common/views/components/SearchFilters/SearchFilters.tsx
@@ -2,9 +2,8 @@ import { useState, FunctionComponent, ReactElement } from 'react';
 import { ParsedUrlQuery } from 'querystring';
 import SearchFiltersDesktop from '../SearchFiltersDesktop/SearchFiltersDesktop';
 import SearchFiltersMobile from '../SearchFiltersMobile/SearchFiltersMobile';
-import { WorksProps } from '../WorksLink/WorksLink';
-import { ImagesProps } from '../ImagesLink/ImagesLink';
 import { LinkProps } from '../../../model/link-props';
+import { Filter } from '../../../services/catalogue/filters';
 
 type Props = {
   query: string;
@@ -16,46 +15,6 @@ type Props = {
 
 export type SearchFiltersSharedProps = Props;
 
-export type DateRangeFilter = {
-  type: 'dateRange';
-  id: string;
-  label: string;
-  to: {
-    key: keyof WorksProps;
-    id: string;
-    value: string | undefined;
-  };
-  from: {
-    key: keyof WorksProps;
-    id: string;
-    value: string | undefined;
-  };
-};
-
-export type CheckboxFilter = {
-  type: 'checkbox';
-  key: keyof WorksProps;
-  id: string;
-  label: string;
-  options: {
-    id: string;
-    value: string;
-    count: number;
-    label: string;
-    selected: boolean;
-  }[];
-};
-
-export type ColorFilter = {
-  type: 'color';
-  key: keyof ImagesProps;
-  id: string;
-  label: string;
-  color: string | undefined;
-};
-
-export type Filter = CheckboxFilter | DateRangeFilter | ColorFilter;
-
 const SearchFilters: FunctionComponent<Props> = ({
   query,
   searchForm,
@@ -63,7 +22,7 @@ const SearchFilters: FunctionComponent<Props> = ({
   filters,
   linkResolver,
 }: Props): ReactElement<Props> => {
-  const [isMobile, setIsMobile] = useState(false);
+  const [isMobile] = useState(false);
 
   const sharedProps: SearchFiltersSharedProps = {
     query,
